+++ packages/integration-sdk-testing/src/jest.ts
@@ -3,8 +3,14 @@ import * as deepmerge from 'deepmerge';
 import {
   Entity,
   ExplicitRelationship,
+  IntegrationInstanceConfig,
+  IntegrationInvocationConfig,
+  IntegrationSpecConfig,
+  IntegrationStepExecutionContext,
   MappedRelationship,
+  Step,
 } from '@jupiterone/integration-sdk-core';
+import { getMatchers } from 'expect/build/jestMatchersObject';
 
 declare global {
   // eslint-disable-next-line @typescript-eslint/no-namespace
@@ -19,6 +25,10 @@ declare global {
       ): R;
 
       toTargetEntities(entities: Entity[]): R;
+
+      toImplementSpec<T extends IntegrationInstanceConfig>(
+        spec: IntegrationSpecConfig<T>,
+      ): R;
     }
   }
 }
@@ -438,10 +448,64 @@ function areKeysDistinct(_keys: string[]): boolean {
   return Array.isArray(_keys) && new Set(_keys).size === _keys.length;
 }
 
+type StepBase<TConfig extends IntegrationInstanceConfig> = Omit<
+  Step<IntegrationStepExecutionContext<TConfig>>,
+  'executionHandler'
+>;
+
+export function toImplementSpec<
+  TConfig extends IntegrationInstanceConfig = IntegrationInstanceConfig
+>(
+  integration: IntegrationInvocationConfig<TConfig>,
+  spec: IntegrationSpecConfig<TConfig>,
+) {
+  const unimplementedSteps: string[] = [];
+
+  const implementedStepsProposed: { [id: string]: StepBase<TConfig> } = {};
+  const implementedStepsActual: { [id: string]: StepBase<TConfig> } = {};
+  const implementedStepsByIdMap: {
+    [id: string]: Step<IntegrationStepExecutionContext<TConfig>>;
+  } = integration.integrationSteps.reduce(
+    (implStepsById, step) => ({ ...implStepsById, [step.id]: step }),
+    {},
+  );
+
+  for (const specStep of spec.integrationSteps) {
+    if (specStep.implemented === false) {
+      unimplementedSteps.push(specStep.id);
+    } else {
+      const stepId = specStep.id;
+      const { implemented, ...specStepBase } = specStep;
+      implementedStepsProposed[stepId] = specStepBase;
+
+      const implementedStep = implementedStepsByIdMap[specStep.id];
+
+      if (implementedStep) {
+        const { executionHandler, ...implementedStepBase } = implementedStep;
+        implementedStepsActual[stepId] = implementedStepBase;
+      }
+    }
+  }
+  if (unimplementedSteps.length > 0) {
+    console.log(
+      {
+        unimplementedSteps,
+      },
+      'Spec steps marked as `implemented: false`',
+    );
+  }
+  return getMatchers().toMatchObject.call(
+    expect,
+    implementedStepsActual,
+    implementedStepsProposed,
+  );
+}
+
 export function registerMatchers(expect: jest.Expect) {
   expect.extend({
     toMatchGraphObjectSchema,
     toMatchDirectRelationshipSchema,
     toTargetEntities,
+    toImplementSpec,
   });
 }